feat: Change translator back to class and create dictionary module
