+++ braille.txt
@@ -4,12 +4,6 @@
 0.
 .0
 ..
-0.
-0.
-0.
-0.
-0.
-0.
-0.
+00
 .0
-0.
+00
