+++ message.txt
@@ -1,1 +1,1 @@
-hello
+hey
